+++ xposed-stub/src/main/java/de/robv/android/xposed/IXposedHookZygoteInit.java
@@ -1,0 +1,10 @@
+package de.robv.android.xposed;
+
+public interface IXposedHookZygoteInit extends IXposedMod {
+    void initZygote(StartupParam startupParam) throws Throwable;
+
+    final class StartupParam {
+        public String modulePath;
+        public boolean startsSystemServer;
+    }
+}